+++ src/tests/query.rs
@@ -1457,6 +1457,44 @@ async fn enum_values(api: &mut dyn TestApi) -> crate::Result<()> {
     Ok(())
 }
 
+#[test_each_connector(tags("postgresql"))]
+#[cfg(all(feature = "json", feature = "postgresql"))]
+async fn row_to_json_normal(api: &mut dyn TestApi) -> crate::Result<()> {
+    let cte = Select::default()
+        .value(val!("hello_world").alias("toto"))
+        .into_cte("one");
+    let select = Select::from_table("one").value(row_to_json("one", false)).with(cte);
+    let result = api.conn().select(select).await?;
+
+    assert_eq!(
+        Value::Json(Some(serde_json::json!({
+            "toto": "hello_world"
+        }))),
+        result.into_single().unwrap()[0]
+    );
+
+    Ok(())
+}
+
+#[test_each_connector(tags("postgresql"))]
+#[cfg(all(feature = "json", feature = "postgresql"))]
+async fn row_to_json_pretty(api: &mut dyn TestApi) -> crate::Result<()> {
+    let cte = Select::default()
+        .value(val!("hello_world").alias("toto"))
+        .into_cte("one");
+    let select = Select::from_table("one").value(row_to_json("one", true)).with(cte);
+    let result = api.conn().select(select).await?;
+
+    assert_eq!(
+        Value::Json(Some(serde_json::json!({
+            "toto": "hello_world"
+        }))),
+        result.into_single().unwrap()[0]
+    );
+
+    Ok(())
+}
+
 #[test_each_connector(ignore("mysql"))]
 async fn single_common_table_expression(api: &mut dyn TestApi) -> crate::Result<()> {
     let cte = Select::default()